requirements.txt

# to read API key from env file 
python-dotenv

# for Gemini
google-generativeai

# Web API
fastapi 
uvicorn
websockets